Issue #4437: Modified UnusedImportsCheckTest.java and moved its input files to the unusedimports subdirectory
