+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/imports/UnusedImportsCheckTest.java
@@ -36,13 +36,13 @@ public class UnusedImportsCheckTest extends BaseCheckTestSupport {
     @Override
     protected String getPath(String filename) throws IOException {
         return super.getPath("checks" + File.separator
-                + "imports" + File.separator + filename);
+                + "imports" + File.separator + "unusedimports" + File.separator + filename);
     }
 
     @Override
     protected String getNonCompilablePath(String filename) throws IOException {
         return super.getNonCompilablePath("checks" + File.separator
-                + "imports" + File.separator + filename);
+                + "imports" + File.separator + "unusedimports" + File.separator + filename);
     }
 
     @Test
@@ -108,21 +108,21 @@ public class UnusedImportsCheckTest extends BaseCheckTestSupport {
     public void testProcessJavadocWithLinkTag() throws Exception {
         final DefaultConfiguration checkConfig = createCheckConfig(UnusedImportsCheck.class);
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
-        verify(checkConfig, getPath("InputUnusedImportWithValueTag.java"), expected);
+        verify(checkConfig, getPath("InputUnusedImportsWithValueTag.java"), expected);
     }
 
     @Test
     public void testAnnotations() throws Exception {
         final DefaultConfiguration checkConfig = createCheckConfig(UnusedImportsCheck.class);
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
-        verify(checkConfig, getNonCompilablePath("package-info.java"), expected);
+        verify(checkConfig, getNonCompilablePath("InputUnusedImportsAnnotations.java"), expected);
     }
 
     @Test
     public void testBug() throws Exception {
         final DefaultConfiguration checkConfig = createCheckConfig(UnusedImportsCheck.class);
         final String[] expected = CommonUtils.EMPTY_STRING_ARRAY;
-        verify(checkConfig, getPath("InputImportBug.java"), expected);
+        verify(checkConfig, getPath("InputUnusedImportsBug.java"), expected);
     }
 
     @Test
@@ -182,7 +182,7 @@ public class UnusedImportsCheckTest extends BaseCheckTestSupport {
             "5:8: " + getCheckMessage(MSG_KEY, "java.util.Arrays"),
             "6:8: " + getCheckMessage(MSG_KEY, "java.lang.String"),
         };
-        verify(checkConfig, getNonCompilablePath("InputRedundantImport_UnnamedPackage.java"),
+        verify(checkConfig, getNonCompilablePath("InputUnusedImportsFileInUnnamedPackage.java"),
             expected);
     }
 
@@ -202,6 +202,6 @@ public class UnusedImportsCheckTest extends BaseCheckTestSupport {
             "12:8: " + getCheckMessage(MSG_KEY, "java.lang.Float"),
             "13:8: " + getCheckMessage(MSG_KEY, "java.lang.Short"),
         };
-        verify(checkConfig, getPath("InputUnusedImportFromJavaLang.java"), expected);
+        verify(checkConfig, getPath("InputUnusedImportsFromJavaLang.java"), expected);
     }
 }
