+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/imports/unusedimports/InputUnusedImportsAnnotations.java
@@ -1,5 +1,5 @@
 ////////////////////////////////////////////////////////////////////////////////
-// Annotated package definition to check if import is declared unused
+// this file content is based on package-info.java file structure
 // Created: 2005
 ////////////////////////////////////////////////////////////////////////////////
 @MyAnnotation
