+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/imports/unusedimports/InputUnusedImportsFileInUnnamedPackage.java
@@ -1,0 +1,13 @@
+// no package
+
+import java.util.List;
+import java.util.List;
+import java.util.Arrays;
+import java.lang.String;
+import static java.lang.Math.PI;
+
+public class InputUnusedImportsFileInUnnamedPackage
+{
+    public static double pi=PI;
+    public List myList;
+}
